+++ laboratory/gradle-plugin/src/main/kotlin/io/mehow/laboratory/gradle/FeatureFlagOptionInput.kt
@@ -8,6 +8,6 @@ internal class FeatureFlagOptionInput(private val name: String, private val isDe
   fun toModel() = FeatureFlagOption(name, isDefault)
 
   internal companion object {
-    private const val serialVersionUID = 0L
+    private const val serialVersionUID = 1L
   }
 }