+++ laboratory/gradle-plugin/src/main/kotlin/io/mehow/laboratory/gradle/PackageNameProvider.kt
@@ -13,6 +13,6 @@ internal class PackageNameProvider(private val delegate: PackageNameProvider? = 
   }
 
   internal companion object {
-    private const val serialVersionUID = 0L
+    private const val serialVersionUID = 1L
   }
 }