+++ laboratory/gradle-plugin/src/main/kotlin/io/mehow/laboratory/gradle/SourcedFeatureStorageInput.kt
@@ -27,6 +27,6 @@ internal constructor(packageNameProvider: PackageNameProvider) : Serializable {
     )
 
   internal companion object {
-    private const val serialVersionUID = 0L
+    private const val serialVersionUID = 1L
   }
 }