+++ laboratory/inspector/src/main/kotlin/io/mehow/laboratory/inspector/FeatureUiModel.kt
@@ -1,7 +1,6 @@
 package io.mehow.laboratory.inspector
 
 import io.mehow.laboratory.Feature
-import io.mehow.laboratory.supervisorOption
 
 internal data class FeatureUiModel(
   val type: Class<out Feature<*>>,
@@ -11,7 +10,6 @@ internal data class FeatureUiModel(
   val sources: List<OptionUiModel>,
   val deprecationAlignment: DeprecationAlignment?,
   val deprecationPhenotype: DeprecationPhenotype?,
-  val supervisorOption: Feature<*>?,
 ) {
   val hasMultipleSources = sources.size > 1
 
@@ -19,9 +17,7 @@ internal data class FeatureUiModel(
     sources.firstOrNull(OptionUiModel::isSelected)?.option?.name?.equals("Local", ignoreCase = true)
       ?: true
 
-  private val isSupervised = type.supervisorOption == supervisorOption
-
-  val isEnabled = isCurrentSourceLocal && isSupervised
+  val isEnabled = isCurrentSourceLocal
 
   companion object {
     private val firstAlignmentOrdinal = DeprecationAlignment.values().first()